Move "add_term_to_batch" to Terms (and copy to beaconing (just to avoid errors))
