+++ app/tests/unit_tests/test_analyzer_terms.py
@@ -40,6 +40,16 @@ class TestTermsAnalyzer(unittest.TestCase):
         self.test_settings.restore_default_configuration_path()
         self.test_es.restore_es()
 
+    @staticmethod
+    def _preperate_data_terms():
+        eval_terms_array = defaultdict()
+        # "random" value
+        aggregator_value = "key"
+        target_value = "test"
+        observations = {'a': 1, 'test': 'ok'}
+        doc = {'source': 'this', 'target': 12}
+        return eval_terms_array, aggregator_value, target_value, observations, doc
+
     def _test_whitelist_batch_document_not_process_all(self):  # TODO FIX with new whitelist system
         self.test_settings.change_configuration_path("/app/tests/unit_tests/files/terms_test_with_whitelist.conf")
         analyzer = TermsAnalyzer("terms_dummy_test")
@@ -148,3 +158,52 @@ class TestTermsAnalyzer(unittest.TestCase):
             if "outliers" in doc['_source']:
                 nbr_outliers += 1
         self.assertEqual(nbr_outliers, len(all_doc))
+
+    def test_add_term_to_batch_empty(self):
+        eval_terms_array = defaultdict()
+        aggregator_value = ""
+        target_value = ""
+        observations = {}
+        doc = {}
+        # Create expected result
+        expected_eval_terms = defaultdict()
+        expected_eval_terms[aggregator_value] = defaultdict(list)
+        expected_eval_terms[aggregator_value]["targets"] = [target_value]
+        expected_eval_terms[aggregator_value]["observations"] = [{}]
+        expected_eval_terms[aggregator_value]["raw_docs"] = [{}]
+
+        self.assertEqual(TermsAnalyzer.add_term_to_batch(eval_terms_array, aggregator_value, target_value, observations,
+                                                         doc), expected_eval_terms)
+
+    def test_add_term_to_batch_no_modification(self):
+        eval_terms_array, aggregator_value, target_value, observations, doc = self._preperate_data_terms()
+        # Create expected result
+        expected_eval_terms = defaultdict()
+        expected_eval_terms[aggregator_value] = defaultdict(list)
+        expected_eval_terms[aggregator_value]["targets"] = [target_value]
+        expected_eval_terms[aggregator_value]["observations"] = [observations]
+        expected_eval_terms[aggregator_value]["raw_docs"] = [doc]
+
+        self.assertEqual(TermsAnalyzer.add_term_to_batch(eval_terms_array, aggregator_value, target_value, observations,
+                                                         doc), expected_eval_terms)
+
+    def test_add_term_to_batch_concerv_extra_value(self):
+        eval_terms_array, aggregator_value, target_value, observations, doc = self._preperate_data_terms()
+        # Add extra value:
+        eval_terms_array["newKey"] = defaultdict(list)
+        eval_terms_array["newKey2"] = "empty"
+        eval_terms_array[aggregator_value] = defaultdict(list)
+        eval_terms_array[aggregator_value]["targets"] = [target_value]
+        eval_terms_array[aggregator_value]["test"] = 12
+        # Create expected result
+        expected_eval_terms = defaultdict()
+        expected_eval_terms["newKey"] = defaultdict(list)
+        expected_eval_terms["newKey2"] = "empty"
+        expected_eval_terms[aggregator_value] = defaultdict(list)
+        expected_eval_terms[aggregator_value]["targets"] = [target_value, target_value]
+        expected_eval_terms[aggregator_value]["observations"] = [observations]
+        expected_eval_terms[aggregator_value]["raw_docs"] = [doc]
+        expected_eval_terms[aggregator_value]["test"] = 12
+
+        self.assertEqual(TermsAnalyzer.add_term_to_batch(eval_terms_array, aggregator_value, target_value, observations,
+                                                         doc), expected_eval_terms)
